Rename util helper classes
