+++ embc-app/ClientApp/src/app/shared/utils/environment-utils.ts
