+++ embc-app/ClientApp/src/app/shared/utils/index.ts
@@ -1,2 +1,2 @@
-export * from './environmentUtils';
-export * from './stateUtils';
+export * from './environment-utils';
+export * from './state-utils';
