+++ embc-app/ClientApp/src/app/shared/utils/state-utils.ts
